pyproject.toml
[project]
name = "tau-evo"
version = "0.1.0"
description = "Self-evolving LLM agents via reflective prompt and tool-schema patching on tau2-bench"
requires-python = ">=3.11"
dependencies = [
    "tau2",
    "openai>=1.0",
    "python-dotenv",
    "rich>=13.0",
    "textual>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "sse-starlette>=2.0",
]

[tool.uv.sources]
tau2 = { path = "tau2-bench", editable = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["evo"]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = ["E501"]

[tool.ruff]
line-length = 100
